fix: missing models
